Change cache getting and setting
The gateway api caches requests made to the other internal api endpoints
but by default cachecontrol caches these requests by their url. We want
to include a user's auth token in the cache key (for private requests)
so that we can cache by user.

When reading from the cache we now always use the authorization header
to build the cache key in conjunction with the request.url. When
actually reading from the backend store we rely on redis' mget method to
request both versions of the key (with and without the hash of the auth
header) and return the first value we find. We are biased towards keys
that do NOT include the auth hash but in practice public/private should
be a binary switch and so there should never be an issue of a less
preferred key being returned.

When writing to the cache we examine the response when creating a cache
key to determine if the response is private or public so we can set an
appropriate cache key